feat: Migrate from script tags to ES modules
This commit refactors the project to use ES modules by updating the HTML to use `importmap` and removing Babel. It also reorganizes the script loading and integrates Tailwind CSS for styling.

Additionally, it introduces a new start overlay and curtain effect to enhance the user experience before the application begins. The core JavaScript logic is also restructured for better maintainability.